requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
idna==3.3
lxml==4.6.3
multidict==5.2.0
multitasking==0.0.9
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
slackclient==2.9.3
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
yfinance==0.1.63
